aptiv-s32g: remove HVP from atf and linux
Issue: LIN1022-2393

Since we update nxp-s32g to BSP35, the HVP patch for atf will cause
do_patch failure, let's remove this patch to fix this issue and wait for
the conflicts resolving.

And remove the linux HVP patch to fix the boot failure which caused by
the removal of the atf HVP patch.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>