aptiv-s32g: improve strategy of reporting error
Issue: LIN1022-3546,LIN1022-3547

When building normal aptiv-s32g BSP with HSE firmware, the variable
HSE_SEC_ENABLED is set as 1. In this way, fatal error is reported and
building is stopped. In order to avoid affecting normal aptiv-s32g BSP
building, report issue only if autosar secure boot and nxp secure boot
are enabled at the same time.
In additional, improve the conditional expression code so that the
redundant code can be dropped.

Signed-off-by: Meng Li <Meng.Li@windriver.com>