aptiv-s32g: save autosar image offset used to config parallel secure boot
Issue: LINUXEXEC-26092

When configing the parallel secure boot, the atf image is brought up by BOOTROM,
because autosar image can't bring up atf image any more. In order to verify the
autosar image, it needs to install a SMR(secure memory region) for it. So, it needs
to save the autosar image offset at 0x1110 from the beginning of bootloader image.

Signed-off-by: Meng Li <Meng.Li@windriver.com>