aptiv: conf: remove hvp feature from default MACHINE_FEATURES
Issue: LIN1024-2475

Now the aptiv-s32g bsp has been updated to compatible with SDK BSP41
release, and the related atf, uboot, kernel, etc. codes are all updated,
but the hvp related patches are still in old status, so there will be
errors when try to apply hvp related patches, like following:

  ERROR: atf-s32g-2.10-r0 do_patch: Applying patch '0005-aptiv-s32g-atf-add-xrdc.patch' on target directory '.../atf-s32g/2.10/git'
  BBHandledException("Applying '0005-aptiv-s32g-atf-add-xrdc.patch' failed:
  stdout: checking file fdts/s32cc.dtsi
  can't find file to patch at input line 47

Currently, the hvp feature is not default needed, so remove it from the
MACHINE_FEATURES list by default to fix the above build error.

And the hvp feature update needs virtualization team support and will be
updated in future.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>